Refactored TypeNameDifferenceResolverUtils away.
