+++ src/NUnitFramework/framework/Internal/TypeNameDifferenceResolver.cs
@@ -32,16 +32,6 @@ namespace NUnit.Framework.Internal
     /// </summary>
     public class TypeNameDifferenceResolver
     {
-        private TypeNameDifferenceResolverUtils _utils;
-
-        /// <summary>
-        /// Instantiate a <see cref="TypeNameDifferenceResolver"/>.
-        /// </summary>
-        public TypeNameDifferenceResolver()
-        {
-            _utils = new TypeNameDifferenceResolverUtils();
-        }
-
         /// <summary>
         /// Gets the shortened type name difference between <paramref name="expected"/> and <paramref name="actual"/>.
         /// </summary>
@@ -63,25 +53,25 @@ namespace NUnit.Framework.Internal
         /// <param name="actualTypeShortened">Output of the unique type name for actual object.</param>
         public void ResolveTypeNameDifference(Type expected, Type actual, out string expectedTypeShortened, out string actualTypeShortened)
         {
-            if (_utils.IsTypeGeneric(expected) && _utils.IsTypeGeneric(actual))
+            if (IsTypeGeneric(expected) && IsTypeGeneric(actual))
             {
                 string shortenedGenericTypeExpected, shortenedGenericTypeActual;
-                _utils.GetShortenedGenericTypes(expected, actual, out shortenedGenericTypeExpected, out shortenedGenericTypeActual);
+                GetShortenedGenericTypes(expected, actual, out shortenedGenericTypeExpected, out shortenedGenericTypeActual);
 
                 List<string> shortenedParamsExpected, shortenedParamsActual;
                 GetShortenedGenericParams(expected, actual, out shortenedParamsExpected, out shortenedParamsActual);
 
-                expectedTypeShortened = _utils.ReconstructGenericTypeName(shortenedGenericTypeExpected, shortenedParamsExpected);
-                actualTypeShortened = _utils.ReconstructGenericTypeName(shortenedGenericTypeActual, shortenedParamsActual);
+                expectedTypeShortened = ReconstructGenericTypeName(shortenedGenericTypeExpected, shortenedParamsExpected);
+                actualTypeShortened = ReconstructGenericTypeName(shortenedGenericTypeActual, shortenedParamsActual);
             }
-            else if (_utils.IsTypeGeneric(expected) || _utils.IsTypeGeneric(actual))
+            else if (IsTypeGeneric(expected) || IsTypeGeneric(actual))
             {
-                expectedTypeShortened = _utils.FullyShortenTypeName(expected);
-                actualTypeShortened = _utils.FullyShortenTypeName(actual);
+                expectedTypeShortened = FullyShortenTypeName(expected);
+                actualTypeShortened = FullyShortenTypeName(actual);
             }
             else
             {
-                _utils.ShortenTypeNames(expected, actual, out expectedTypeShortened, out actualTypeShortened);
+                ShortenTypeNames(expected, actual, out expectedTypeShortened, out actualTypeShortened);
             }
         }
 
@@ -114,13 +104,131 @@ namespace NUnit.Framework.Internal
 
             foreach (Type genericParamRemaining in templateParamsExpected)
             {
-                shortenedParamsExpected.Add(_utils.FullyShortenTypeName(genericParamRemaining));
+                shortenedParamsExpected.Add(FullyShortenTypeName(genericParamRemaining));
             }
 
             foreach (Type genericParamRemaining in templateParamsActual)
             {
-                shortenedParamsActual.Add(_utils.FullyShortenTypeName(genericParamRemaining));
+                shortenedParamsActual.Add(FullyShortenTypeName(genericParamRemaining));
             }
+        }
+
+        /// <summary>
+        /// Obtain a shortened name of the given <see cref="Type"/>.
+        /// </summary>
+        public string FullyShortenTypeName(Type GenericType)
+        {
+            if (IsTypeGeneric(GenericType))
+            {
+                string genericType = GenericType.GetGenericTypeDefinition().Name;
+
+                List<Type> genericParams = new List<Type>(GenericType.GetGenericArguments());
+                List<string> shortenedGenericParams = new List<string>();
+                genericParams.ForEach(x => shortenedGenericParams.Add(FullyShortenTypeName(x)));
+
+                return ReconstructGenericTypeName(genericType, shortenedGenericParams);
+            }
+            else
+            {
+                return GenericType.Name;
+            }
+        }
+
+        /// <summary>
+        /// Shorten the given <see cref="Type"/> names by only including the relevant differing namespaces/types, if they differ.
+        /// </summary>
+        /// <param name="expectedType">The expected <see cref="Type"/>.</param>
+        /// <param name="actualType">The actual <see cref="Type"/>.</param>
+        /// <param name="expectedTypeShortened">The shortened expected <see cref="Type"/> name.</param>
+        /// <param name="actualTypeShortened">The shortened actual <see cref="Type"/> name.</param>
+        public void ShortenTypeNames(Type expectedType, Type actualType, out string expectedTypeShortened, out string actualTypeShortened)
+        {
+            string[] expectedOriginalType = expectedType.FullName.Split('.');
+            string[] actualOriginalType = actualType.FullName.Split('.');
+
+            bool diffDetected = false;
+            int actualStart = 0, expectStart = 0;
+            for (int expectLen = expectedOriginalType.Length - 1, actualLen = actualOriginalType.Length - 1;
+                expectLen >= 0 && actualLen >= 0;
+                expectLen--, actualLen--)
+            {
+                if (expectedOriginalType[expectLen] != actualOriginalType[actualLen])
+                {
+                    actualStart = actualLen;
+                    expectStart = expectLen;
+                    diffDetected = true;
+                    break;
+                }
+            }
+
+            if (diffDetected)
+            {
+                expectedTypeShortened = String.Join(".", expectedOriginalType, expectStart, expectedOriginalType.Length - expectStart);
+                actualTypeShortened = String.Join(".", actualOriginalType, actualStart, actualOriginalType.Length - actualStart);
+            }
+            else
+            {
+                expectedTypeShortened = expectedOriginalType[expectedOriginalType.Length - 1];
+                actualTypeShortened = actualOriginalType[actualOriginalType.Length - 1];
+            }
+
+        }
+
+        /// <summary>
+        /// Returns whether or not the <see cref="Type"/> is generic.
+        /// </summary>
+        public bool IsTypeGeneric(Type type)
+        {
+            Guard.ArgumentNotNull(type, nameof(type));
+
+            return type.GetGenericArguments().Length > 0;
+        }
+
+        /// <summary>
+        /// Returns the fully qualified generic <see cref="Type"/> name of a given <see cref="Type"/>.
+        /// </summary>
+        public string GetGenericTypeName(Type type)
+        {
+            Guard.ArgumentNotNull(type, nameof(type));
+
+            if (IsTypeGeneric(type))
+            {
+                Type generic = type.GetGenericTypeDefinition();
+                return generic.FullName;
+            }
+            else
+            {
+                throw new ArgumentException($"The provided {nameof(type)} was not generic");
+            }
+        }
+
+        /// <summary>
+        /// Reconstruct a generic type name using the provided generic type name, and a
+        /// <see cref="List"/> of the template parameters.
+        /// </summary>
+        /// <param name="GenericTypeName">The name of the generic type, including the number of template parameters expected.</param>
+        /// <param name="TemplateParamNames">A <see cref="List"/> of names of the template parameters of the provided generic type.</param>
+        public string ReconstructGenericTypeName(string GenericTypeName, List<string> TemplateParamNames)
+        {
+            return GenericTypeName + "[" + string.Join(",", TemplateParamNames.ToArray()) + "]";
+        }
+
+        /// <summary>
+        /// Obtain the shortened generic <see cref="Type"/> names of the given expected and actual <see cref="Type"/>s.
+        /// </summary>
+        /// <param name="expected">The expected <see cref="Type"/>.</param>
+        /// <param name="actual">The actual <see cref="Type"/>.</param>
+        /// <param name="shortenedGenericNameExpected">The shortened expected generic name.</param>
+        /// <param name="shortenedGenericNameActual">The shortened actual generic name.</param>
+        public void GetShortenedGenericTypes(Type expected, Type actual, out string shortenedGenericNameExpected, out string shortenedGenericNameActual)
+        {
+            Type toplevelGenericExpected = expected.GetGenericTypeDefinition();
+            Type toplevelGenericActual = actual.GetGenericTypeDefinition();
+            ShortenTypeNames(
+                toplevelGenericExpected,
+                toplevelGenericActual,
+                out shortenedGenericNameExpected,
+                out shortenedGenericNameActual);
         }
     }
 }
