+++ src/NUnitFramework/framework/nunit.framework-4.0.csproj
@@ -418,7 +418,6 @@
     <Compile Include="Internal\ThreadUtility.cs" />
     <Compile Include="Internal\TypeHelper.cs" />
     <Compile Include="Internal\TypeNameDifferenceResolver.cs" />
-    <Compile Include="Internal\TypeNameDifferenceResolverUtils.cs" />
     <Compile Include="Internal\TypeWrapper.cs" />
     <Compile Include="Is.cs" />
     <Compile Include="ITestAction.cs" />
