+++ src/NUnitFramework/framework/nunit.framework-4.5.csproj
@@ -107,7 +107,6 @@
     <Compile Include="Internal\TestCaseTimeoutException.cs" />
     <Compile Include="Internal\InvalidPlatformException.cs" />
     <Compile Include="Internal\TypeNameDifferenceResolver.cs" />
-    <Compile Include="Internal\TypeNameDifferenceResolverUtils.cs" />
     <Compile Include="Warn.cs" />
     <Compile Include="Assume.cs" />
     <Compile Include="Attributes\ApartmentAttribute.cs" />
