+++ src/NUnitFramework/framework/nunit.framework-netstandard16.csproj
@@ -426,7 +426,6 @@
     <Compile Include="Internal\ThreadUtility.cs" />
     <Compile Include="Internal\TypeHelper.cs" />
     <Compile Include="Internal\TypeNameDifferenceResolver.cs" />
-    <Compile Include="Internal\TypeNameDifferenceResolverUtils.cs" />
     <Compile Include="Internal\TypeWrapper.cs" />
     <Compile Include="Is.cs" />
     <Compile Include="ITestAction.cs" />
