+++ src/NUnitFramework/tests/nunit.framework.tests-2.0.csproj
@@ -138,7 +138,6 @@
     <Compile Include="Internal\TestNamingTests.cs" />
     <Compile Include="Internal\ThreadUtilityTests.cs" />
     <Compile Include="Internal\TypeHelperTests.cs" />
-    <Compile Include="Internal\TypeNameDifferenceResolverUtilsTests.cs" />
     <Compile Include="Internal\TypeNameDifferenceTests.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Assertions\ArrayEqualsFailureMessageFixture.cs" />
