+++ src/NUnitFramework/tests/nunit.framework.tests-3.5.csproj
@@ -137,7 +137,6 @@
     <Compile Include="Internal\TestNamingTests.cs" />
     <Compile Include="Internal\ThreadUtilityTests.cs" />
     <Compile Include="Internal\TypeHelperTests.cs" />
-    <Compile Include="Internal\TypeNameDifferenceResolverUtilsTests.cs" />
     <Compile Include="Internal\TypeNameDifferenceTests.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Assertions\ArrayEqualsFailureMessageFixture.cs" />
